+++ NEWS.md
@@ -1,3 +1,11 @@
+# tbl.now 0.7.5
+
+* Bumped roxygen to version 8.0.0. This also resulted in updated documentation. 
+* Changed `autoplot()`'s default level to 0.95 
+* Added tests for converters and pillars.
+* Throws warning when converting to `baselinenowcast` if data is 
+`"count-cumulative"`. 
+
 # tbl.now 0.7.3
 
 * Added the `update_now()` function to make it more intuitive to update